Rename ContentLoader component, fix Circle preview, and cleanup unused box.css file
