+++ react/components/Preview/Circle.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { ContentLoader, Box } from './ContentLoader2'
+import { ContentLoader, Rect } from './ContentLoader'
 
 interface Props {
   width: number
@@ -8,7 +8,7 @@ interface Props {
 
 const Circle = ({ width, height }: Props) => (
   <ContentLoader width={width} height={height}>
-    <Box width={width} height={height} borderRadius="100%" />
+    <Rect width={width} height={height} borderRadius="100%" />
   </ContentLoader>
 )
