--- react/components/Preview/box.css
@@ -1,25 +1,0 @@
-@keyframes spin {
-  0% {
-    transform: rotate(0);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
-
-.spin {
-  animation: spin 1s infinite;
-}
-
-@keyframes slide {
-  0% {
-    transform: translate3d(-50%, 0, 0);
-  }
-  100% {
-    transform: translate3d(0, 0, 0);
-  }
-}
-
-.slide {
-  animation: slide 1s infinite linear;
-}
